refactor reporting out of generator into ReportBuilder class
